requirement.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.3.0
PyJWT==2.6.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
